+++ src/main/java/com/JDR/Vacunassist/Repository/ZonaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.JDR.Vacunassist.Model.Zona;
 
 @Repository
-public interface ZonaRepository extends JpaRepository<Zona, Long>{
+public interface ZonaRepository extends JpaRepository<Zona, Integer>{
 
 }
